+++ manual/release-notes.rst
@@ -158,6 +158,8 @@ For a detailed list of changes, please see the file
         output and errors that slipped through the cracks with
         ``setOutputStreams``.
 
+      - A C API is available in :file:`include/qpdf/qpdflogger-c.h`.
+
     - New methods ``insertItemAndGet``, ``appendItemAndGet``,
       ``eraseItemAndGet``, ``replaceKeyAndGet``, and
       ``removeKeyAndGet`` return the newly added or removed object.